backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
Pillow==10.0.1
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
Alembic==1.12.0
gunicorn==21.2.0
# Updated to support image generation
google-generativeai>=0.8.0